refactor: componentized group chat UI for better structure and maintainability
